drivers: net: mqnic: Fix compiler warnings to improve portalibity
- Fix incorrect pointer check in mqnic_platform_module_eeprom_put()
by checking individual array elements instead of the array address.
- Cast dma_addr_t to u64 before right-shifting by 32 bits in
BAR address setup to avoid shift-count-overflow warnings on
32-bit platforms.
- Apply the same fix for all BAR address writes in EQ, TX, and
other relevant code paths.
- These changes silence -Waddress and -Wshift-count-overflow
warnings and improve portability and correctness.

+++ drivers/net/mqnic/mqnic_cq.c
@@ -69,7 +69,7 @@ int mqnic_open_cq(struct mqnic_cq *cq, struct mqnic_eq *eq, int size)
 	// set base address
 	iowrite32((cq->buf_dma_addr & 0xfffff000),
 			cq->hw_addr + MQNIC_CQ_BASE_ADDR_VF_REG + 0);
-	iowrite32(cq->buf_dma_addr >> 32,
+	iowrite32(((u64)cq->buf_dma_addr) >> 32,
 			cq->hw_addr + MQNIC_CQ_BASE_ADDR_VF_REG + 4);
 	// set size
 	iowrite32(MQNIC_CQ_CMD_SET_SIZE | ilog2(cq->size),
